[feat] 이미지 업로드 API 적용 (#80)
