+++ src/api/api.ts
@@ -8,8 +8,8 @@ export const api = axios.create({
   withCredentials: true,
   headers: {
     'Content-Type': 'application/json',
-    Authorization:
-      'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIiwiZXhwIjoxNzcwNDM3NTc4LCJpYXQiOjE3NzAzNTExNjcsImp0aSI6ImQ3YWMxOTExOTMwMDRiZWU4ZTdmMDNjNzE3NWViZDg4IiwidXNlcl9pZCI6MTB9.6yD4w-h05SD-bmSueCB-hXAwSeYkQPjNSyfcVze79H8',
+    // Authorization:
+    //   'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIiwiZXhwIjoxNzcwNDM3NTc4LCJpYXQiOjE3NzAzNTExNjcsImp0aSI6ImQ3YWMxOTExOTMwMDRiZWU4ZTdmMDNjNzE3NWViZDg4IiwidXNlcl9pZCI6MTB9.6yD4w-h05SD-bmSueCB-hXAwSeYkQPjNSyfcVze79H8',
   },
 })
 
@@ -58,43 +58,7 @@ function withAuth(token?: string) {
   if (!token) return {}
   return { Authorization: `Bearer ${token}` }
 }
-/**
- * S3 이미지 업로드를 위한 Presigned URL 발급 요청
- * @param fileName - 업로드할 파일명 (예: "example.png")
- * @returns presigned_url, img_url, key
- */
-export interface PresignedUrlResponse {
-  presigned_url: string
-  img_url: string
-  key: string
-}
-export async function getPresignedUrl(
-  fileName: string
-): Promise<PresignedUrlResponse> {
-  const token = getAccessToken()
-  const res = await api.put<PresignedUrlResponse>(
-    'https://api.ozcodingschool.site/api/v1/qna/answers/presigned-url',
-    { file_name: fileName },
-    { headers: { ...withAuth(token || undefined) } }
-  )
-  return res.data
-}
 
-/**
- * Presigned URL을 사용하여 S3에 파일 업로드
- * @param presignedUrl - 백엔드에서 발급받은 Presigned URL
- * @param file - 업로드할 파일
- */
-export async function uploadToS3(
-  presignedUrl: string,
-  file: File
-): Promise<void> {
-  await axios.put(presignedUrl, file, {
-    headers: {
-      'Content-Type': file.type,
-    },
-  })
-}
 // =============================
 // Community API
 // =============================
